requirements.txt
# ---------------- Core LangChain & Agent Framework ----------------
langchain>=0.2.11
langgraph>=0.2.24
langsmith>=0.1.90
langchain-text-splitters>=0.2.0
langchain-community>=0.4.1
# ---------------- Hugging Face LLM ----------------
transformers>=4.43.0
torch>=2.3.0
accelerate>=0.33.0

# ---------------- Embeddings & RAG ----------------
langchain-huggingface>=0.1.0
sentence-transformers>=2.6.1
rank-bm25>=0.2.2
fastembed>=0.7.4

# ---------------- Vector Database ----------------
qdrant-client>=1.16.2
langchain-qdrant>=1.1.0

# ---------------- Document Processing ----------------
pypdf>=4.2.0

# ---------------- API & Utilities ----------------
requests>=2.32.3
python-dotenv>=1.0.1

# ---------------- UI ----------------
streamlit>=1.36.0

# ---------------- Testing ----------------
pytest>=8.2.2
pytest-mock>=3.14.0
